requirements.txt
langchain-openai
langchain-core
langchain-community
python-dotenv
openai
pydantic
